+++ morpheus-graphql-examples-scotty/src/Server/Mythology/API.hs
@@ -9,7 +9,6 @@ module Server.Mythology.API
   )
 where
 
-
 import Data.Morpheus
   ( App,
     deriveApp,
@@ -58,7 +57,7 @@ data DeityArgs = DeityArgs
   { name :: Text, -- Required Argument
     bornPlace :: Maybe City -- Optional Argument
   }
-  deriving (Generic)
+  deriving (Generic, GQLType)
 
 resolveDeity :: DeityArgs -> ResolverQ e IO Deity
 resolveDeity DeityArgs {name, bornPlace} =