+++ morpheus-graphql-examples-scotty/src/Server/Sophisticated/API.hs
@@ -102,7 +102,7 @@ instance GQLScalar Euro where
           (round (fromIntegral x / 100 :: Double))
           (mod x 100)
       )
-  parseValue _ = Left "invalid currency blue"
+  parseValue _ = Left "invalid currency blue!"
   serialize (Euro x y) = Int (x * 100 + y)
 
 data Channel = USER | ADDRESS
@@ -131,7 +131,7 @@ root =
           querySomeMap = pure $ M.fromList [("robin", 1), ("carl", 2)],
           queryWrapped1 = constRes $ A (0, "some value"),
           queryWrapped2 = pure $ A "",
-          queryFail1 = fail "fail example",
+          queryFail1 = fail "fail example!!",
           queryFail2 = liftEither alwaysFail,
           queryShared =
             pure
@@ -149,13 +149,13 @@ root =
     mutationResolver =
       Mutation
         { mutationCreateUser = resolveCreateUser,
-          mutationCreateAddress = resolveCreateAdress,
-          mutationSetAdress = resolveSetAdress
+          mutationCreateAddress = resolveCreateAddress,
+          mutationSetAddress = resolveSetAddress
         }
     subscriptionResolver =
       Subscription
         { subscriptionNewUser = resolveNewUser,
-          subscriptionNewAddress = const resolveNewAdress
+          subscriptionNewAddress = const resolveNewAddress
         }
 
 -- Resolve QUERY
@@ -180,15 +180,15 @@ resolveCreateUser = do
   liftEither setDBUser
 
 -- Mutation With Event Triggering : sends events to subscription
-resolveCreateAdress :: ResolverM EVENT IO Address
-resolveCreateAdress = do
+resolveCreateAddress :: ResolverM EVENT IO Address
+resolveCreateAddress = do
   requireAuthorized
   publish [addressUpdate]
   lift setDBAddress
 
 -- Mutation Without Event Triggering
-resolveSetAdress :: ResolverM EVENT IO Address
-resolveSetAdress = lift setDBAddress
+resolveSetAddress :: ResolverM EVENT IO Address
+resolveSetAddress = lift setDBAddress
 
 -- Resolve SUBSCRIPTION
 resolveNewUser :: SubscriptionField (ResolverS EVENT IO User)
@@ -198,8 +198,8 @@ resolveNewUser = subscribe USER $ do
   where
     subResolver (Event _ content) = liftEither (getDBUser content)
 
-resolveNewAdress :: SubscriptionField (ResolverS EVENT IO Address)
-resolveNewAdress = subscribe ADDRESS $ do
+resolveNewAddress :: SubscriptionField (ResolverS EVENT IO Address)
+resolveNewAddress = subscribe ADDRESS $ do
   requireAuthorized
   pure subResolver
   where
@@ -235,7 +235,7 @@ getDBUser _ = do
           userEmail = pure email,
           userAddress = const $ lift (getDBAddress (Content 12)),
           userOffice = constRes Nothing,
-          userHome = pure HH,
+          userHome = pure CityIDHH,
           userEntity =
             pure
               [ MyUnionAddress
@@ -250,7 +250,7 @@ getDBUser _ = do
                       userEmail = pure email,
                       userAddress = const $ lift (getDBAddress (Content 12)),
                       userOffice = constRes Nothing,
-                      userHome = pure HH,
+                      userHome = pure CityIDHH,
                       userEntity = pure []
                     }
               ]
@@ -278,7 +278,7 @@ setDBUser = do
         userEmail = pure email,
         userAddress = const $ lift setDBAddress,
         userOffice = constRes Nothing,
-        userHome = pure HH,
+        userHome = pure CityIDHH,
         userEntity = pure []
       }