+++ src/Data/Morpheus/Kind.hs
@@ -12,7 +12,7 @@ module Data.Morpheus.Kind
     INPUT_OBJECT,
     GQL_KIND,
     Context (..),
-    VContext (..),
+    ContextValue (..),
     ResContext (..),
     OUTPUT,
     INPUT,
@@ -42,8 +42,8 @@ data ResContext (kind :: GQL_KIND) (operation :: OperationType) event (m :: * ->
 data Context (kind :: GQL_KIND) a
   = Context
 
-newtype VContext (kind :: GQL_KIND) a = VContext
-  { unVContext :: a
+newtype ContextValue (kind :: GQL_KIND) a = ContextValue
+  { unContextValue :: a
   }
 
 -- | GraphQL Scalar: Int, Float, String, Boolean or any user defined custom Scalar type
@@ -61,12 +61,12 @@ type OUTPUT = 'OUTPUT
 -- | GraphQL input Object and input union
 type INPUT = 'INPUT
 
-{-# DEPRECATED INPUT_OBJECT "use more generalised kind: INPUT" #-}
+{-# DEPRECATED INPUT_OBJECT "use more generalized kind: INPUT" #-}
 
 -- | GraphQL input Object
 type INPUT_OBJECT = 'INPUT
 
-{-# DEPRECATED UNION "use: deriving(GQLType), INPORTANT: only types with <type constructor name><constructor name> will sustain their form, other union constructors will be wrapped inside an new object" #-}
+{-# DEPRECATED UNION "use: deriving(GQLType), IMPORTANT: only types with <type constructor name><constructor name> will sustain their form, other union constructors will be wrapped inside an new object" #-}
 
 -- | GraphQL Union
 type UNION = 'OUTPUT